質問の詳細用のcomponentを作成
